Moved files from Client and renaming
