+++ Turnbased-Game/Models/Packets/Server/PlayerJoinedLobbyPacket.cs
@@ -2,9 +2,9 @@ using Turnbased_Game.Models.Packets.Shared;
 
 namespace Turnbased_Game.Models.Packets.Server;
 
-public class PlayerJoinedLobbyPacket(int playerId, IPlayerProfile profile) : IPackage
+public class PlayerJoinedLobbyPacket(int playerId, PlayerProfile profile) : IPackage
 {
     public byte PacketId => 15;
     public int PlayerId { get; set; } = playerId;
-    public IPlayerProfile Profile{ get; set; } = profile;
+    public PlayerProfile Profile{ get; set; } = profile;
 }
